+++ Data/model.h
@@ -25,7 +25,7 @@ class Model
 public:
     Model();
 
-    size_t addPicket(const Picket& picket);
+    int addPicket(const Picket& picket);
 
     const std::vector<Node>& getNodes() const;
     const std::vector<Picket>& getPickets() const;
