+++ Services/picketgenerator.cpp
@@ -9,7 +9,7 @@ void PicketGenerator::generatePickets(Model *model)
 {
     auto& nodes = model->getNodes();
 
-    for(size_t i = 0; i < nodes.size(); ++i)
+    for(int i = 0; i < nodes.size(); ++i)
     {
         model->addPicket(Picket { &nodes[i] });
     }
